+++ src/estd/port/arduino/iostream.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "istream.h"
+#include "ostream.h"
+
+namespace estd {
+
+// UNTESTED, but compiles
+using arduino_iostream = 
+    detail::basic_ostream<estd::arduino_ostreambuf, arduino_istream>;
+
+}