+++ src/estd/port/iostream.h
@@ -3,9 +3,6 @@
 #include "istream.h"
 #include "ostream.h"
 
-namespace estd {
-
 #ifdef ESTD_ARDUINO
+#include "arduino/iostream.h"
 #endif
-
-}